requirements.txt
torch
torchvision
matplotlib
scikit-learn
tqdm
numpy
pillow
streamlit
kagglehub
pandas